Update shuffling function

+++ src/state/config.rs
@@ -1,13 +1,13 @@
 pub struct Config {
-    pub attester_count: u32,
-    pub max_validators: u32
+    pub attester_count: u64,
+    pub max_validators: u64
 }
 
 impl Config {
     pub fn standard() -> Self {
         Self {
             attester_count: 32,
-            max_validators: 2u32.pow(24)
+            max_validators: 2u64.pow(24)
         }
     }
 }
